examples/agent_basics/requirements.txt
asana==5.0.0
openai==1.10.0
python-dotenv==0.13.0
